+++ src/game/server/gamecontext_kz.cpp
@@ -533,14 +533,14 @@ void CGameContext::SendDiscordChatMessage(int ClientID, const char* msg)
 	m_pHttp->Run(pDiscord);
 }
 
-void CGameContext::SendDiscordRecordMessage(int ClientID, float Time, float PrevTime)
+void CGameContext::SendDiscordRecordMessage(int ClientID, double Time, double PrevTime)
 {
 	char aPayload[4048];
 	char aStatsStr[4000];
 	char aStr[500];
 	aStr[0] = '\0';
 
-	str_format(aStr, sizeof(aStr),"New record on map %s by %s: %d minute(s) %f second(s)!!!", Server()->GetMapName(), Server()->ClientName(ClientID), (int)Time / 60, Time - ((int)Time / 60 * 60));
+	str_format(aStr, sizeof(aStr),"New record on map %s by %s: %d minute(s) %.6f second(s)!!!", Server()->GetMapName(), Server()->ClientName(ClientID), (int)Time / 60, Time - ((int)Time / 60 * 60));
 
 	str_format(
 		aPayload,
